+++ src/main/resources/templates/fragments/header.html
@@ -25,7 +25,7 @@
                 <ul class="dropdown-menu text-small">
                     <li><a class="dropdown-item" href="#">Nuevo Projecto</a></li>
                     <li><a class="dropdown-item" href="#">Configuracion</a></li>
-                    <li><a class="dropdown-item" th:href="@{/perfil}" th:if="${session.usuariosession != null}"> Mi Perfil </a></li>
+                    <li><a class="dropdown-item" th:href="@{usuario/modificar_perfil_usuario}" th:if="${session.usuariosession != null}"> Modificar Perfil </a></li>
                     <li><hr class="dropdown-divider"></li>
                     <li><a class="dropdown-item" href="/logout">Cerrar Sesion</a></li>
                 </ul>
